NPC interaction transitions with buttons
