+++ ARPG-CSE5912-LTS/Assets/ImportedAssets/SmallImports/TextMesh Pro/Fonts/BOD_R SDF.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5628a06387c340b46a9c188e2da55510
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
