--- ARPG-CSE5912-LTS/Assets/Scripts/RPG System Components/DialogueSystem/DialogueActivator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2b800744c6208e64db3d76adb6edd79d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
