--- ARPG-CSE5912-LTS/Assets/Scripts/RPG System Components/DialogueSystem/DialogueObject.cs
@@ -1,15 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-[CreateAssetMenu(menuName = "Dialogue/DialogueObject")]
-public class DialogueObject : ScriptableObject
-{
-    [SerializeField] [TextArea] private string[] dialogue;
-    [SerializeField] private Response[] responses;
-
-    public string[] Dialogue => dialogue;
-
-    public bool HasResponses => Responses != null && Responses.Length > 0;
-
-    public Response[] Responses => responses;
-}
