--- ARPG-CSE5912-LTS/Assets/Scripts/RPG System Components/DialogueSystem/DialogueObject.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: af4ab72d280a62f4c973857668d1c1b9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
