--- ARPG-CSE5912-LTS/Assets/Scripts/RPG System Components/DialogueSystem/DialogueUI.cs
@@ -1,59 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using TMPro;
-public class DialogueUI : MonoBehaviour
-{
-    [SerializeField] private GameObject dialogueBox;
-    [SerializeField] private TMP_Text textLabel;
-    
-    public bool IsOpen { get; private set; }
-
-    private ResponseHandler responseHandler;
-
-    private TypeWriterEffect typeWriterEffect;
-
-    private void Start()
-    {
-        typeWriterEffect = GetComponent<TypeWriterEffect>();
-        responseHandler = GetComponent<ResponseHandler>();
-        CloseDialogueBox();
-    }
-
-    public void ShowDialogue(DialogueObject dialogueObject)
-    {
-        IsOpen = true;
-        dialogueBox.SetActive(true);
-        StartCoroutine(routine:StepThroughDialouge(dialogueObject));
-    }
-
-    private IEnumerator StepThroughDialouge(DialogueObject dialogueObject)
-    {
-        for (int i = 0; i < dialogueObject.Dialogue.Length; i++)
-        {
-            string dialogue = dialogueObject.Dialogue[i];
-            yield return typeWriterEffect.Run(dialogue, textLabel);
-
-            if (i == dialogueObject.Dialogue.Length - 1 && dialogueObject.HasResponses) break;
-
-            yield return new WaitUntil(() => Input.GetKeyDown(KeyCode.Mouse0));
-        }
-
-        if (dialogueObject.HasResponses)
-        {
-            responseHandler.ShowReponses(dialogueObject.Responses);
-        }
-        else
-        {
-            CloseDialogueBox();
-        }
-    }
-
-    private void CloseDialogueBox()
-    {
-        IsOpen = false;
-        dialogueBox.SetActive(false);
-        textLabel.text = string.Empty;
-    }
-
-}
