--- ARPG-CSE5912-LTS/Assets/Scripts/RPG System Components/DialogueSystem/IInteractable.cs
@@ -1,9 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public interface IInteractable
-{
-    void Interact(Player player);
-
-}
