--- ARPG-CSE5912-LTS/Assets/Scripts/RPG System Components/DialogueSystem/IInteractable.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6cb8efeb96ef6fe4d8671de44478c2af
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
