--- ARPG-CSE5912-LTS/Assets/Scripts/RPG System Components/DialogueSystem/Response.cs
@@ -1,14 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-[System.Serializable]
-
-public class Response 
-{
-    [SerializeField] private string responseText;
-    [SerializeField] private DialogueObject dialogueObject;
-
-    public string ResponseText => responseText;
-
-    public DialogueObject DialogueObject => dialogueObject;
-}
