--- ARPG-CSE5912-LTS/Assets/Scripts/RPG System Components/DialogueSystem/Response.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 75786c54df9631f45b48e2cd683e7958
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
