--- ARPG-CSE5912-LTS/Assets/Scripts/RPG System Components/DialogueSystem/ResponseHandler.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4baa7fbc02e73ed45aa1696076a84344
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
