+++ ARPG-CSE5912-LTS/Assets/Scripts/RPG System Components/NPCInteractions/Dialogue.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d12c5a8ad1369914ea65c38614a01872
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
