+++ ARPG-CSE5912-LTS/Assets/Scripts/RPG System Components/NPCInteractions/NPC.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 003e6b952f5584f44a733eee68d24584
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
