+++ ARPG-CSE5912-LTS/Assets/Scripts/RPG System Components/NPCInteractions/NPC/NPC.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d97670a93e3e22943aa5966f040e0afb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
