+++ ARPG-CSE5912-LTS/Assets/Scripts/RPG System Components/NPCInteractions/NPC/NPCManager.cs
@@ -1,0 +1,11 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public abstract class NPCManager : MonoBehaviour
+{
+    protected abstract bool Interactable();
+
+    public virtual void Interact() {  }
+
+}
