+++ ARPG-CSE5912-LTS/Assets/Scripts/RPG System Components/NPCInteractions/NPC/NPCManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5c7450596e377af47b8aa473a734076a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
