--- ARPG-CSE5912-LTS/Assets/Scripts/RPG System Components/NPCInteractions/NPCManager.cs
@@ -1,11 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public abstract class NPCManager : MonoBehaviour
-{
-    protected abstract bool Interactable();
-
-    public virtual void Interact() {  }
-
-}
