+++ ARPG-CSE5912-LTS/Assets/Scripts/RPG System Components/NPCInteractions/OlderSystem.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1c3a2b3af1b01e1498636109ba366334
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
