+++ ARPG-CSE5912-LTS/Assets/Scripts/RPG System Components/NPCInteractions/OlderSystem/DialogueUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a4cdc8376833f2f4294ab9250198e187
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
