+++ ARPG-CSE5912-LTS/Assets/Scripts/RPG System Components/NPCInteractions/OlderSystem/IInteractable.cs
@@ -1,0 +1,9 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public interface IInteractable
+{
+    void Interact(Player player);
+
+}
