+++ ARPG-CSE5912-LTS/Assets/Scripts/RPG System Components/NPCInteractions/OlderSystem/Response.cs
@@ -1,0 +1,14 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+[System.Serializable]
+
+public class Response 
+{
+    [SerializeField] private string responseText;
+    [SerializeField] private DialogueObject dialogueObject;
+
+    public string ResponseText => responseText;
+
+    public DialogueObject DialogueObject => dialogueObject;
+}
