+++ ARPG-CSE5912-LTS/Assets/Scripts/RPG System Components/NPCInteractions/OlderSystem/ResponseHandler.cs
@@ -1,0 +1,55 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using TMPro;
+using UnityEngine.UI;
+
+public class ResponseHandler : MonoBehaviour
+{
+    [SerializeField] private RectTransform responseBox;
+    [SerializeField] private RectTransform responseButtonTemplate;
+    [SerializeField] private RectTransform responseContainer;
+
+    private DialogueUI dialogueUI;
+
+    List<GameObject> tempResponseButtons = new List<GameObject>();
+
+    private void Start()
+    {
+        dialogueUI = GetComponent<DialogueUI>();
+    }
+
+    public void ShowReponses(Response[] responses)
+    {
+        float responseBoxHeight = 0;
+        foreach (Response response in responses)
+        {
+            GameObject responseButton = Instantiate(responseButtonTemplate.gameObject, responseContainer);
+            responseButton.gameObject.SetActive(true);
+            responseButton.GetComponent<TMP_Text>().text = response.ResponseText;
+            responseButton.GetComponent<Button>().onClick.AddListener(call: () => OnPickedResponse(response));
+
+            tempResponseButtons.Add(responseButton);
+
+            responseBoxHeight += responseButtonTemplate.sizeDelta.y;
+
+        }
+
+        responseBox.sizeDelta = new Vector2(responseBox.sizeDelta.x, y: responseBoxHeight);
+        responseBox.gameObject.SetActive(true);
+    }
+
+    private void OnPickedResponse(Response response)
+    {
+        Debug.Log("clicked on response");
+        responseBox.gameObject.SetActive(false);
+        
+        foreach (GameObject button in tempResponseButtons)
+        {
+            Destroy(button);
+        }
+        tempResponseButtons.Clear();
+
+        dialogueUI.ShowDialogue(response.DialogueObject);
+    }
+}
