+++ ARPG-CSE5912-LTS/Assets/Scripts/RPG System Components/NPCInteractions/OlderSystem/TypeWriterEffect.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8dfcab811e69e62448b8f023c1309ec8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
